third_party/blink/web_tests/editing/pasteboard/paste-into-table-cell-expected.txt

This tests pasting into a table cell.
To manually run the test, copy the text below and paste it into the first text cell after the non-breaking space.
The content should be pasted into the first cell, not between two tds.

"Drag me" should appear inside the first td:
| <table>
|   width="400"
|   <tbody>
|     <tr>
|       <td>
|         " Drag me<#selection-caret>"
|       <td>
|         " "
|     <tr>
|       <td>
|         " "
|       <td>
|         " "
